mali_kbase: replace 0-length array with flexible array in kutf_alloc_entry
Linux must have all the "fake" flexible array members replaced with
actual C99 flexible array members so that the programmer's intent can
actually be represented in an unambiguous way (see
-fstrict-flex-arrays). Replace the 0-length array with flexible
array - it requires no special handling, simply removing the "0" in the
array declaration is sufficient. Everything that could go wrong with
this conversion becomes a compile-time failure and we're out of them.

Updates for size calculations are not necessary for this struct.

Change-Id: Ic56715c9b403fbfecacadb4370a10c3e5e006bd5
Bug: 292213903
Signed-off-by: Tudor Ambarus <tudordana@google.com>
Link: https://people.kernel.org/kees/bounded-flexible-arrays-in-c
Signed-off-by: Will McVicker <willmcvicker@google.com>