mali_kbase: replace 1-element array with flexible array for firmware_trace_buffer
With "-fstrict-flex-arrays=3" enabled, UBSAN_BOUNDS no longer pretends
1-element arrays are unbounded. Walking name will trigger a warning, so
make it a proper flexible array. Replace the 1-element array with a
true flexible array and update the allocation size.

Keep the original name for the flex array name and don't consider
padding as this structure is declared in a .c file and no other code
base can be using it.

Change-Id: Ie1db264bf999a034b7286baf6f189956fd830dc8
Bug: 292213903
Signed-off-by: Tudor Ambarus <tudordana@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>