mali_kbase: replace 1-element array with flexible array for base_external_resource
With "-fstrict-flex-arrays=3" enabled, UBSAN_BOUNDS no longer pretends
1-element arrays are unbounded. Walking ext_res will trigger a warning,
so make it a proper flexible array. Add a union to keep the struct size
identical for firmware in case anything was depending on the old size.
While here, use struct_size() to compute the overall structure size, it
avoids arithmetic overflows as the result saturates to at most SIZE_MAX.

Keep the original name for the flex array name as grepping through the
android sources revealed a single user which isn't impacted by the
flex_array change:
./vendor/arm/mali/valhall/cmar/mali_cmar_gpu.h

Change-Id: I695bc74b415009c31cb6e5cb6b28204cd257ed60
Bug: 292213903
Signed-off-by: Tudor Ambarus <tudordana@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>