treewide: fix $(srctree)/$(src)/ becomes $(src) (v6.10 compatibility)
Linux commit [1] changes $(src) to always point to the source tree.
Therefore, using $(srctree)/$(src) becomes incorrect, as $(srctree)
points to the kernel source tree and compilation of this driver fails
due to the compiler failing to find the driver's header files.

Here specifically, $(src) already points to the (absolute) source tree
of this module, and prepending $(srctree), the kernel source tree, can
not be of any benefit anymore, just causing confusion.

Update the code accordingly.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=b1992c3772e6 [1]
      ("kbuild: use $(src) instead of $(srctree)/$(src) for source
      directory")
Test: TH
Bug: 354917765
Change-Id: I3649311acccff6431a34ef89cad82fc64581cfe3
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>