treewide: plug memleak: case 2.2.1: for_each_available_child_of_node() (v6.12 prep)
Case 2.2.1 from below bug:

    If returning or breaking early from within
    for_each_available_child_of_node(), a reference to the current
    device_node remains and needs to be dropped, either right away, or
    later if used past that loop to avoid a memleak.

In this case, the np is not needed past the loop and code can be
simplified by switching to for_each_available_child_of_node_scoped()
which avoids the need for manual cleanup in early exits from the loop.

This commit plugs memleaks in error code paths.

Test: TH
Bug: 370679843
Change-Id: If76d370a1af0d856b673793254515f87562e4a29
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>